backend/requirements.txt
# Backend CMEP - Python dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.46
asyncmy==0.2.9
pymysql==1.1.1
alembic==1.14.1

# Auth
bcrypt==4.2.1
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
aiosqlite==0.20.0
anyio==4.8.0
